requirements.txt
# Phase 2: Resume Parsing & Text Preprocessing
# NLP Foundation Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Resume Parsing
pdfplumber==0.10.3
python-docx==1.1.0

# NLP Processing
spacy==3.7.2
# Run after install: python -m spacy download en_core_web_sm

# Text Processing
regex==2023.12.25

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
